nyc_taxi_benchmark/requirements.txt
google-cloud-storage
requests

